Rework view bindings
- make bindings saver be adding more parameter checks
- rename parameter to fit Android ones
