+++ arch/src/main/java/knaufdan/android/arch/databinding/view/StringToIntTranslator.kt
@@ -1,0 +1,18 @@
+package knaufdan.android.arch.databinding.view
+
+import androidx.databinding.InverseMethod
+
+object StringToIntTranslator {
+    @JvmStatic
+    @InverseMethod("fromIntToString")
+    fun fromStringToInt(value: String): Int? =
+        try {
+            Integer.parseInt(value)
+        } catch (e: NumberFormatException) {
+            null
+        }
+
+    @JvmStatic
+    fun fromIntToString(value: Int?): String =
+        value?.toString() ?: ""
+}
